server/requirements.txt
openai
python-dotenv
Flask==3.0.0
flask
flask_cors
gunicorn==20.1.0
Werkzeug==3.0.1